chore: restructure package and edit file paths in demo notebooks
